+++ backend/src/Enum/ComicStatus.php
@@ -7,6 +7,7 @@ namespace App\Enum;
 enum ComicStatus: string
 {
     case BUYING = 'buying';
+    case DOWNLOADING = 'downloading';
     case FINISHED = 'finished';
     case STOPPED = 'stopped';
     case WISHLIST = 'wishlist';
@@ -15,6 +16,7 @@ enum ComicStatus: string
     {
         return match ($this) {
             self::BUYING => 'En cours d\'achat',
+            self::DOWNLOADING => 'En cours de téléchargement',
             self::FINISHED => 'Terminée',
             self::STOPPED => 'Arrêtée',
             self::WISHLIST => 'Liste de souhaits',